Fixes Issue 2830. Fixes Issue 2963. Fixes Issue 2964.
git-svn-id: https://mobicents.googlecode.com/svn/trunk/servers/jain-slee@21027 bf0df8d0-2c1f-0410-b170-bd30377b63dc

+++ core/common/src/main/java/org/mobicents/slee/container/management/jmx/MobicentsManagementMBean.java
@@ -34,13 +34,13 @@ public interface MobicentsManagementMBean extends ServiceMBean {
 	 * 
 	 * @return
 	 */
-	public double getEntitiesRemovalDelay();
+	public int getEntitiesRemovalDelay();
 
 	/**
 	 * 
 	 * @param entitiesRemovalDelay
 	 */
-	public void setEntitiesRemovalDelay(double entitiesRemovalDelay);
+	public void setEntitiesRemovalDelay(int entitiesRemovalDelay);
 		
 	/**
 	 *
